# api/openapi.yaml
openapi: 3.0.3
info:
  title: boxincloud API
  version: 0.1.0-dev
  description: |
    Bibliothèque de BD, comics et mangas sur stockage objet.

    Contrat de l'API boxincloud.

    **Ce fichier est la source de vérité.** Le serveur Go, le client TypeScript
    du web et le client Dart de l'application Flutter en sont générés
    (`make generate`). Toute évolution de l'API commence ici — jamais par le
    code, sinon les clients divergent.

    Un test de contrat (`internal/httpapi/contract_test.go`) valide chaque
    réponse du serveur réel contre ce schéma : une divergence entre le contrat
    et l'implémentation fait échouer la CI.

    ## Erreurs

    Les erreurs suivent la [RFC 9457](https://www.rfc-editor.org/rfc/rfc9457)
    (`application/problem+json`). Le champ `type` est stable : les clients
    peuvent s'appuyer dessus pour brancher leur comportement. Deux valeurs
    méritent un traitement particulier :

    - `.../token-expired` — le jeton d'accès a expiré, rafraîchir puis rejouer ;
    - `.../session-revoked` — la session est révoquée, redemander une connexion
      complète (une réutilisation de refresh token a été détectée).

    ## Pagination

    Pagination par curseur. Le client renvoie `nextCursor` tel quel dans le
    paramètre `cursor` ; il ne doit jamais le construire ni l'interpréter.
    L'absence de `nextCursor` signifie « dernière page ».

    ## Versionnement

    Le préfixe `/api/v1` change uniquement en cas de rupture. Les ajouts
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /libraries:
    post:
      operationId: createLibrary
      summary: Créer une bibliothèque sur un backend existant
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, backendId]
              properties:
                name: { type: string }
                backendId: { type: string, format: uuid }
                kind:
                  type: string
                  description: "Défaut : comics."
                rootPrefix:
                  type: string
                  description: |
                    Préfixe sous lequel la bibliothèque vit dans le backend.
                    Vide pour la racine.
      responses:
        "201":
          description: Bibliothèque créée
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LibraryDetail" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"

    get:
      operationId: listLibraries
      summary: Bibliothèques visibles
      description: |
        Seules les bibliothèques auxquelles le compte a accès sont retournées.
        Une bibliothèque sans restriction est visible de tous.
      tags: [catalog]
      responses:
        "200":
          description: Bibliothèques
          content:
            application/json:
              schema:
                type: object
                required: [libraries]
                properties:
                  libraries:
                    type: array
                    items: { $ref: "#/components/schemas/Library" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /folders:
    get:
      operationId: listFolders
      summary: Arborescence des dossiers
      description: |
        Retourne les dossiers à plat, triés de sorte qu'un parent précède
        toujours ses enfants — le client bâtit l'arbre en une passe, sans
        récursion.

        Les compteurs sont cumulatifs : un dossier affiche le total de sa
        branche, ce qu'attend quelqu'un qui replie un nœud.

        Les dossiers existent en base, ils ne sont plus seulement déduits des
        clés d'objet. Un dossier créé à la main apparaît donc même vide.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/LibraryId"
      responses:
        "200":
          description: Dossiers
          content:
            application/json:
              schema:
                type: object
                required: [folders]
                properties:
                  folders:
                    type: array
                    items: { $ref: "#/components/schemas/Folder" }
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createFolder
      summary: Créer un dossier
      description: |
        Rien n'est écrit dans le stockage : un magasin d'objets n'a pas de
        répertoires, seulement des clés. Le dossier existe donc d'abord dans
        boxincloud, et prendra corps au premier fichier déposé.

        Les dossiers intermédiaires manquants sont créés en passant.
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [libraryId, path]
              properties:
                libraryId: { type: string, format: uuid }
                path:
                  type: string
                  description: |
                    Chemin complet relatif au préfixe de la bibliothèque. Les
                    remontées « .. » sont neutralisées.
      responses:
        "201":
          description: Dossier créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Folder" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Le dossier, ou l'un de ses parents, est en lecture seule
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /folders/path:
    put:
      operationId: relocateFolder
      summary: Renommer ou déplacer un dossier
      description: |
        Renommer et déplacer sont le même geste : le dossier d'un album découle
        de la clé de son objet, si bien que changer le dossier renomme chacun
        des objets qu'il contient — par copie côté serveur quand le backend le
        permet.

        L'opération peut donc durer. Elle est refusée au-delà de deux mille
        albums dans la branche : au-delà, mieux vaut la découper que la subir.

        Un échec en cours de route laisse la branche à moitié déplacée. C'est
        volontaire : défaire des déplacements réussis multiplierait les
        occasions de perdre un fichier, alors qu'un état mixte reste
        entièrement lisible, chaque album pointant la clé où il se trouve.
      tags: [catalog]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [libraryId, path, newPath]
              properties:
                libraryId: { type: string, format: uuid }
                path: { type: string }
                newPath:
                  type: string
                  description: |
                    Nouveau chemin complet. Déplacer un dossier dans sa propre
                    descendance est refusé.
      responses:
        "200":
          description: Dossier après déplacement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Folder" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Le dossier, ou l'un de ses parents, est en lecture seule
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /libraries/{libraryId}/folders:
    delete:
      operationId: deleteFolder
      summary: Supprimer un dossier
      description: |
        Trois degrés, du plus sûr au plus définitif. Sans paramètre, un dossier
        qui contient encore des albums est refusé en 409 : supprimer une branche
        entière ne doit jamais être un geste distrait.

        `deleteComics` retire les albums du catalogue en laissant les fichiers.
        `deleteFiles` les efface aussi, et implique le premier.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/LibraryIdPath"
        - name: path
          in: query
          required: true
          schema: { type: string }
        - name: deleteComics
          in: query
          description: Retirer aussi les albums du catalogue.
          schema: { type: boolean }
        - name: deleteFiles
          in: query
          description: Effacer aussi les fichiers du stockage. Irréversible.
          schema: { type: boolean }
      responses:
        "200":
          description: Dossier supprimé
          content:
            application/json:
              schema:
                type: object
                required: [removedComics]
                properties:
                  removedComics: { type: integer }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: |
            Le dossier contient encore des albums, ou il est en lecture seule.
            Le champ `type` distingue les deux cas.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /comics/{comicId}/folder:
    put:
      operationId: moveComic
      summary: Ranger un album dans un autre dossier
      description: |
        Le dossier d'un album n'est pas une propriété du catalogue : il découle
        de la clé de l'objet dans le backend. Le déplacer renomme donc l'objet,
        par copie côté serveur quand le backend le permet — les octets ne
        transitent jamais par boxincloud.

        Une destination déjà occupée par un fichier du même nom est refusée
        plutôt qu'écrasée.
      tags: [catalog]
      parameters:
        - $ref: "#/components/parameters/ComicId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [folder]
              properties:
                folder:
                  type: string
                  description: |
                    Dossier de destination, relatif au préfixe de la
                    bibliothèque. Vide pour la racine. Les remontées « .. » sont
                    neutralisées.
      responses:
        "200":
          description: Dossier réel après déplacement
          content:
            application/json:
              schema:
                type: object
                required: [folderPath]
                properties:
                  folderPath: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Jeton d'accès obtenu via `POST /auth/login`. Durée de vie courte
        (15 min par défaut) ; à renouveler avec le refresh token.
  parameters:
    ComicId:
      name: comicId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    LibraryIdPath:
      name: libraryId
      in: path
      required: true
      schema: { type: string, format: uuid }
    LibraryId:
      name: libraryId
      in: query
      schema: { type: string, format: uuid }
      description: Restreint à une bibliothèque. Omis, toutes les bibliothèques visibles.
...
  schemas:
...
    Library:
      type: object
      required: [id, name, kind, comicCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind:
          type: string
          enum: [comic, manga, book, mixed]
        comicCount: { type: integer, format: int32 }
...
    LibraryDetail:
      type: object
      required: [id, backendId, name, kind, rootPrefix, comicCount]
      properties:
        id: { type: string, format: uuid }
        backendId: { type: string, format: uuid }
        name: { type: string }
        kind: { type: string }
        rootPrefix: { type: string }
        comicCount: { type: integer, format: int32 }
...
    Folder:
      type: object
      required: [id, libraryId, path, name, depth, comicCount, explicit, readOnly, hasCode, unlocked]
      properties:
        id: { type: string, format: uuid }
        libraryId: { type: string, format: uuid }
        path:
          type: string
          description: Chemin complet, relatif au préfixe. Vide pour la racine.
        name:
          type: string
          description: Dernier segment du chemin — ce qui s'affiche dans l'arbre.
        depth: { type: integer, description: "0 à la racine." }
        comicCount:
          type: integer
          description: Albums de ce dossier et de ses sous-dossiers.
        explicit:
          type: boolean
          description: |
            Vrai pour un dossier créé à la main, qui survit au fait d'être vide.
            Faux pour un dossier constaté par le parcours, élagué dès qu'il ne
            contient plus rien.
        readOnly:
          type: boolean
          description: Protégé contre les modifications. Ne masque rien.
        hasCode:
          type: boolean
          description: |
            Masqué par un code d'accès. Un dossier masqué non déverrouillé
            n'apparaît pas du tout dans cette liste — ce drapeau ne peut donc
            être vrai que sur un dossier déverrouillé, ou pour un administrateur
            qui vient de poser le code.
        unlocked:
          type: boolean
          description: Le code a été saisi et n'a pas expiré.
...
    Problem:
      type: object
      description: |
        Erreur au format RFC 9457, servie avec le type MIME
        `application/problem+json`.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Identifiant stable de la catégorie d'erreur.
        title: { type: string }
        status: { type: integer }
        detail:
          type: string
          description: Explication de ce cas précis, affichable à l'utilisateur.
        instance:
          type: string
          description: |
            Identifiant de la requête. À inclure dans un rapport de bug pour
            retrouver la trace côté serveur.
        errors:
          type: object
          additionalProperties: { type: string }
          description: Champs invalides, pour les erreurs de validation.
  responses:
...
    Unauthorized:
      description: Jeton absent, expiré ou invalide
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: Droits insuffisants
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: Ressource introuvable
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationFailed:
      description: Validation échouée
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...